Change the lengths back to shorts.

+++ apps/s_client.c
@@ -546,8 +546,8 @@ static int next_proto_cb(SSL *s, unsigned char **out, unsigned char *outlen, con
 	}
 # endif  /* ndef OPENSSL_NO_NEXTPROTONEG */
 
-static int serverinfo_cli_cb(SSL* s, unsigned int ext_type,
-												     unsigned char* in, unsigned int inlen, 
+static int serverinfo_cli_cb(SSL* s, unsigned short ext_type,
+												     unsigned char* in, unsigned short inlen, 
 												     int* al, void* arg)
 	{
 	char pem_name[100];
